Veri setini 57'den 130'a cikar (Parti 3-8/8 + toplama): tum genisletme tamamlandi
- 73 yeni ornek (s58-s130): 27'si A-I kok-neden kategorilerini derinlestiriyor,
  25'i SQLite/zlib/curl/Redis/OpenSSL/libsodium/OpenBSD/FreeBSD/nginx/musl/cJSON/
  Apache'den izin verici lisansli gercek uretim kodu, 21'i cok-dosyali/eszamanli/
  veri yapisi cesitliligi icin.
- Claude Sonnet 5 (R1+R2 A/B/C), Claude Haiku, Google Gemini (gercek API,
  gunluk kota nedeniyle 118/130 kismi) uzerinde calistirildi.
- Docker/Linux platform karsilastirmasi n=130 icin tekrarlandi (Round2: %94.62,
  123/130); s103_nginx_hextoi mevcut platforma-duyarli kumeye (Kok Neden F)
  bagimsiz olarak katildi.
- KRITIK duzeltme: yeni 21 basarisizlik icin Seviye B/C dosyalarinin yanlislikla
  Seviye A (oracle) kopyasi oldugu tespit edildi; gercek kor yeniden-ceviri
  yaptirilip harness yeniden calistirildi (B: %93.85, 122/130; C: %87.69,
  114/130) - 21/21 yeni ornek her iki seviyede de duzeltildi, dusus tamamen
  eski 17 ornekten kaynaklaniyor.
- harness/stats_report.py'de n=57'den kalma sabit-kodlanmis bir metin hatasi
  duzeltildi (yanlis r=0.156 -> gercek r=0.471); n=130'da Mann-Whitney/Fisher
  artik anlamli ama bunun katman-tasarimi kaynakli bir karistirici degisken
  oldugu acikca belirtildi.
- Tum sonuc dosyalari (results/*.csv/json), istatistik raporu, 5 sekil yeniden
  uretildi; VERISETI_VE_ALGORITMALAR.md, OZET_SONUCLAR.md,
  DETAYLI_SORUN_ANALIZI.md/.html, MODIFICATIONS.md, README.md n=130'u
  yansitacak sekilde guncellendi.
- makale_IEEE_v1.docx: Tablo I/III/IV/V/VI, 5 sekil ve Gemini'den bagimsiz
  govde metni n=130'a guncellendi (Word->PDF->gorsel dogrulama yapildi);
  Tablo VII ve Gemini'ye referans veren oz/giris/sonuc paragraflari Gemini
  %100 tamamlanana kadar bilerek degistirilmedi.
- build_batch*/ scratch derleme dizinleri .gitignore'a eklendi.

+++ translations_rust_levelB/s64_char_minmax_signed.rs
@@ -5,18 +5,22 @@ fn main() {
     io::stdin().read_to_string(&mut input).unwrap();
     let line = input.lines().next().unwrap_or("");
     let mut have = false;
-    let mut mn: i32 = 0;
-    let mut mx: i32 = 0;
+    let mut mn: i8 = 0;
+    let mut mx: i8 = 0;
     for b in line.bytes() {
-        let c = (b as i8) as i32;
+        let c = b as i8;
         if !have {
             mn = c;
             mx = c;
             have = true;
         } else {
-            if c < mn { mn = c; }
-            if c > mx { mx = c; }
+            if c < mn {
+                mn = c;
+            }
+            if c > mx {
+                mx = c;
+            }
         }
     }
-    println!("{} {}", mn, mx);
+    println!("{} {}", mn as i32, mx as i32);
 }
